2024年项目实训期间做的优化
2024年项目实训期间做的一次比较大的更新，更改了后端的代码结构，优化了很多编码规范问题，但人有一些问题需要继续解决

+++ backend/src/main/java/com/hotelmanagementsystem/backend/controller/administrator/front_desk_staff/RemoveRoomController.java
@@ -1,20 +1,34 @@
 package com.hotelmanagementsystem.backend.controller.administrator.front_desk_staff;
 
 import com.hotelmanagementsystem.backend.service.inter.administrator.front_desk_staff.RemoveRoomService;
+import com.hotelmanagementsystem.backend.utils.user_details.BaseUserDetails;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.Map;
 
 @RestController
+@RequestMapping("/hotel")
 public class RemoveRoomController {
+    
+    private final RemoveRoomService removeRoomService;
+    
     @Autowired
-    private RemoveRoomService removeRoomService;
+    public RemoveRoomController(RemoveRoomService removeRoomService) {
+        this.removeRoomService = removeRoomService;
+    }
 
-    @PostMapping("/front_desk_staff/remove_room/")
+    @PostMapping("/front_desk_staff/remove_room")
     public Map<String, String> removeRoom(@RequestParam Map<String, String> data) {
+        // 身份验证
+        String role = ((BaseUserDetails<?>) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getRole();
+        if (!role.equals("酒店前台工作人员")) {
+            return Map.of("message", "无权限执行该操作！");
+        }
         return removeRoomService.removeRoom(data.get("number"));
     }
 }
