+++ backend/src/main/java/com/hotelmanagementsystem/backend/controller/administrator/hr_staff/RemoveStaffController.java
@@ -2,20 +2,34 @@ package com.hotelmanagementsystem.backend.controller.administrator.hr_staff;
 
 import com.hotelmanagementsystem.backend.service.inter.administrator.hr_staff.AddStaffService;
 import com.hotelmanagementsystem.backend.service.inter.administrator.hr_staff.RemoveStaffService;
+import com.hotelmanagementsystem.backend.utils.user_details.BaseUserDetails;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.util.Map;
 
 @RestController
+@RequestMapping("/hotel")
 public class RemoveStaffController {
+    
+    private final RemoveStaffService removeStaffService;
+    
     @Autowired
-    private RemoveStaffService removeStaffService;
+    public RemoveStaffController(RemoveStaffService removeStaffService) {
+        this.removeStaffService = removeStaffService;
+    }
 
-    @PostMapping("/hr_staff/remove_staff/")  //要修改数据库，所以用PostMapping
+    @PostMapping("/hr_staff/remove_staff")
     public Map<String, String> removeStaff(@RequestParam Map<String, String> data) {
+        // 身份验证
+        String role = ((BaseUserDetails<?>) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getRole();
+        if (!role.equals("人事管理员")) {
+            return Map.of("message", "无权限执行该操作！");
+        }
         return removeStaffService.removeStaff(data);
     }
 }
